dev/experimental-utils/pdfcat/pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "pdfcat"
version = "0.1.0"
description = "A PDF viewer for Ghostty terminal using Kitty graphics protocol"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "PyMuPDF>=1.23.0",
    "click>=8.1.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
pdfcat = "pdfcat.cli:main"

[tool.setuptools]
packages = ["pdfcat"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
